chore: move to cli & support mf restart
