+++ packages/core/src/cli/build.ts
@@ -1,15 +1,12 @@
 import { type RsbuildInstance, createRsbuild } from '@rsbuild/core';
 import { composeRsbuildEnvironments, pruneEnvironments } from '../config';
-import { onBeforeRestartServer, watchFilesForRestart } from '../restart';
 import type { RslibConfig } from '../types/config';
 import type { BuildOptions } from './commands';
-import { loadRslibConfig } from './init';
+import { onBeforeRestartServer } from './restart';
 
 export async function build(
   config: RslibConfig,
-  options: Pick<BuildOptions, 'lib' | 'watch'> & {
-    configFilePath?: string;
-  } = {},
+  options: Pick<BuildOptions, 'lib' | 'watch'> = {},
 ): Promise<RsbuildInstance> {
   const environments = await composeRsbuildEnvironments(config);
   const rsbuildInstance = await createRsbuild({
@@ -22,24 +19,8 @@ export async function build(
     watch: options.watch,
   });
 
-  if (options?.watch) {
-    const files: string[] = [];
-
-    if (options.configFilePath) {
-      files.push(options.configFilePath);
-    }
-
+  if (options.watch) {
     onBeforeRestartServer(buildInstance.close);
-
-    watchFilesForRestart(files, async () => {
-      const { content: rslibConfig, filePath: configFilePath } =
-        await loadRslibConfig(options);
-
-      await build(rslibConfig, {
-        configFilePath,
-        ...options,
-      });
-    });
   } else {
     await buildInstance.close();
   }
