+++ packages/core/src/cli/commands.ts
@@ -5,6 +5,7 @@ import { build } from './build';
 import { loadRslibConfig } from './init';
 import { inspect } from './inspect';
 import { startMFDevServer } from './mf';
+import { watchFilesForRestart } from './restart';
 
 export type CommonOptions = {
   root?: string;
@@ -62,13 +63,20 @@ export function runCli(): void {
     .description('build the library for production')
     .action(async (options: BuildOptions) => {
       try {
-        const { content: rslibConfig, filePath: configFilePath } =
-          await loadRslibConfig(options);
-        await build(rslibConfig, {
-          configFilePath,
-          lib: options.lib,
-          watch: options.watch,
-        });
+        const cliBuild = async () => {
+          const { content: rslibConfig, filePath } =
+            await loadRslibConfig(options);
+
+          await build(rslibConfig, options);
+
+          if (options?.watch) {
+            watchFilesForRestart([filePath], async () => {
+              await cliBuild();
+            });
+          }
+        };
+
+        await cliBuild();
       } catch (err) {
         logger.error('Failed to build.');
         logger.error(err);
@@ -110,9 +118,18 @@ export function runCli(): void {
     .description('start Rsbuild dev server of Module Federation format')
     .action(async (options: CommonOptions) => {
       try {
-        const { content: rslibConfig } = await loadRslibConfig(options);
-        // TODO: support lib option in mf dev server
-        await startMFDevServer(rslibConfig);
+        const mfDev = async () => {
+          const { content: rslibConfig, filePath } =
+            await loadRslibConfig(options);
+          // TODO: support lib option in mf dev server
+          await startMFDevServer(rslibConfig);
+
+          watchFilesForRestart([filePath], async () => {
+            await mfDev();
+          });
+        };
+
+        await mfDev();
       } catch (err) {
         logger.error('Failed to start mf dev.');
         logger.error(err);
